fix: Refactored namespaces globally
